face-service/requirements.txt
face_recognition==1.3.0
flask==3.1.0
numpy>=1.24.0
requests>=2.31.0
Pillow>=10.0.0
scikit-learn>=1.3.0
gunicorn>=21.2.0
